refacor: MatrixMain, Readers, Writers

+++ src/MatrixMul/Extensions/IntExtension.cs
@@ -3,5 +3,5 @@ namespace MatrixMul.Extensions;
 public static class IntExtension
 {
     public static int IntPow(this int number, int powValue)
-        => (int) Math.Pow((double)number, (double)powValue);
+        => (int) Math.Pow(number, powValue);
 }
